[pointfog] Centralize database calls
